adding tv evaluate

+++ src/data/dataset.py
@@ -214,7 +214,7 @@ class Strand:
                     left = None
 
             names = ["mean", "stdv", "length", "start"]
-            D = {n: np.array(v, dtype=np.float64) for n, v in zip(
+            D = {n: np.array(v, dtype=np.float32) for n, v in zip(
                 names, np.array(self.signal_bc).T[1:])}
             D["seq"] = np.array(self.signal_bc).T[0]
